aix.h (ASM_GENERATE_INTERNAL_LABEL): Fix format specifier.
	* rs6000/aix.h (ASM_GENERATE_INTERNAL_LABEL): Fix format specifier.

	* xcoffout.c (assign_type_number): Constify.
	(xcoffout_source_file): Add static prototype.  Don't needlessly
	cast away const-ness.

From-SVN: r34866

+++ gcc/config/rs6000/aix.h
@@ -462,7 +462,7 @@ toc_section ()						\
    This is suitable for output with `assemble_name'.  */
 
 #define ASM_GENERATE_INTERNAL_LABEL(LABEL,PREFIX,NUM)	\
-  sprintf (LABEL, "*%s..%d", PREFIX, NUM)
+  sprintf (LABEL, "*%s..%ld", (PREFIX), (long)(NUM))
 
 /* This is how to output an assembler line to define N characters starting
    at P to FILE.  */
